Add general OAuth2/OIDC option
Add option to provide a token url directly, instaed of an AAD tenant.
Also provide an (optional) resource parameter.
